Huge updated, haven't done proper commits-see desc
Spent the last 3 days better understanding webpack

1st learning point:
I reconfigured my webpack config as it was full of bloat loaders. Took a lot of time to really think about how each file actually gets bundled and built to be used in the browser.

2nd learning point: After painstakingly trying to replicate my src directory in my dist directory, I was able to this with conversations with ChatGPT (awesome learning tool) However, ChatGPT did not provide me with the desired outcome from the get go-- I went on a goose chase downloading a copy plugin which actually exacerbated the issue at hand. What finally helped is when I realized the issue was with the asset/resource management and I was able to ask for a proper solution by referencing the filename property like such:         filename: (pathData) => {
            const folderPath = path.dirname(pathData.filename).replace(/^src\//, "");
            return `${folderPath}/[name][ext]`;
          },

3rd learning point.
I learned how to read through an entire array of images and to use this array as a reference point for my images to load. This using the webpack require context and creating a function called importimage which examines the length of the files that exist in the specified image/subdirectory and returns that image directory as an array of strings that reference the location of those images.

Working point 1:

With all of this I was finally able to actually generate a working slot machine that randomly chooses weapon images to put to the slot machine.